Remove conditional inclussion of subscribers
This reverts the change made in 13093048790b097816ce39adf29d0a4137e6909f,
as we are facing issues on sites that have transmogrify.dexterity installed after migration.

The change was included to avoid an AttributeError while migrating content.

We'll need to find a different workaround for this in the future in case needed.